Defaultmodel based on pydantic
